healthCheck implemented
