+++ nee-ctx-web-ktor/src/test/kotlin/pl/setblack/nee/ctx/web/BasicAuthProviderTest.kt
@@ -1,4 +1,4 @@
-package pl.seetblack.nee.ctx.web
+package pl.setblack.nee.ctx.web
 
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.be
