+++ nee-ctx-web-ktor/src/test/kotlin/pl/setblack/nee/ctx/web/KtorThreadingModelTest.kt
@@ -1,10 +1,9 @@
-package pl.seetblack.nee.ctx.web
+package pl.setblack.nee.ctx.web
 
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.longs.shouldBeGreaterThan
 import io.kotest.matchers.longs.shouldBeLessThan
 import io.ktor.application.Application
-import io.ktor.application.ApplicationCall
 import io.ktor.application.call
 import io.ktor.http.HttpMethod
 import io.ktor.response.respondText
@@ -13,55 +12,13 @@ import io.ktor.routing.routing
 import io.ktor.server.testing.TestApplicationEngine
 import io.ktor.server.testing.createTestEnvironment
 import io.ktor.server.testing.handleRequest
-import kotlinx.coroutines.IO_PARALLELISM_PROPERTY_NAME
 import kotlinx.coroutines.newFixedThreadPoolContext
 import pl.setblack.nee.Nee
-import pl.setblack.nee.ctx.web.BaseWebContext
-import pl.setblack.nee.ctx.web.DefaultErrorHandler
-import pl.setblack.nee.ctx.web.JDBCBasedWebContext
-import pl.setblack.nee.ctx.web.WebContext
-import pl.setblack.nee.ctx.web.WebContextProvider
-import pl.setblack.nee.effects.Out
-import pl.setblack.nee.effects.async.ECProvider
-import pl.setblack.nee.effects.async.ExecutionContextProvider
-import pl.setblack.nee.effects.async.ExecutorExecutionContext
-import pl.setblack.nee.effects.jdbc.JDBCProvider
-import pl.setblack.nee.effects.security.SecurityCtx
-import pl.setblack.nee.effects.security.SecurityError
-import pl.setblack.nee.effects.security.SecurityProvider
-import pl.setblack.nee.effects.tx.TxConnection
-import pl.setblack.nee.effects.tx.TxProvider
-import pl.setblack.nee.effects.utils.invalid
-import pl.setblack.nee.security.User
-import pl.setblack.nee.security.UserRole
-import java.sql.Connection
+import pl.setblack.nee.ctx.web.support.EmptyTestContext
 import java.util.concurrent.CountDownLatch
 import java.util.concurrent.Executors
 
 fun Application.slowApp() {
-    val myTxProvider = object : TxProvider<Connection, JDBCProvider> {
-        override fun getConnection(): TxConnection<Connection> =
-            invalid()
-
-        override fun setConnectionState(newState: TxConnection<Connection>): JDBCProvider =
-            invalid()
-    }
-
-    val noSecurity = object : SecurityProvider<User, UserRole> {
-        override fun getSecurityContext(): Out<SecurityError, SecurityCtx<User, UserRole>> =
-            invalid()
-    }
-
-    val serverExecutor = Executors.newFixedThreadPool(KtorThreadingModelTest.reqs)
-    val ec = ExecutorExecutionContext(serverExecutor)
-
-    val provider  = object : BaseWebContext<Connection, JDBCProvider>() {
-        override val txProvider = myTxProvider
-        override fun authProvider(call: ApplicationCall): SecurityProvider<User, UserRole> =
-            noSecurity
-
-        override val executionContextProvider = ECProvider(ec)
-    }
 
     routing {
         get("/slow") {
@@ -70,8 +27,8 @@ fun Application.slowApp() {
             call.respondText { "ok" }
         }
         get("/fast") {
-            val wc = provider.create(call)
-            val result = Nee.constP(provider.effects().async) {
+            val wc = EmptyTestContext.contexProvider.create(call)
+            val result = Nee.constP(EmptyTestContext.contexProvider.effects().async) {
                 Thread.sleep(100)
                 "ok"
             }.perform(wc)(Unit)
